use shared BackendError for both uniffi and wasm
