+++ crates/yomikiri-backend/src/ffi/mod.rs
@@ -1,3 +1,2 @@
 mod backend;
-mod error;
 mod update_dict;
